pyproject.toml
[project]
name = "movie-rating-system"
version = "0.1.0"
description = ""
authors = [
    { name = "Reyhaneh Hajihashemi and Sanaz Aparviz", email = "reyhanehhashemi171@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = ["fastapi (>=0.128.0,<0.129.0)", "uvicorn[standard] (>=0.40.0,<0.41.0)", "sqlalchemy (>=2.0.45,<3.0.0)", "psycopg2-binary (>=2.9.11,<3.0.0)", "alembic (>=1.17.2,<2.0.0)", "pydantic-settings (>=2.12.0,<3.0.0)", "python-dotenv (>=1.2.1,<2.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
